Refacted views to extend from a base CircuitView for code reuse
